fix: rename router to routes
